Set old decorator test pending
@petercrona - which the new decorator interface, which is implemented
as a plugin, this test cannot be written in this way anymore, because
the internals of entityStore and queryCache are hidden.

How important is the behavior here? I suppose the goal here is to test
if cross entity functionality is correct (invalidation).
Ideally we would not need to touch the internals of queryCache to test
this here anyway.
Is this functionality tested elsewhere so that we can just remove this
test here, or should we try to re-write in a different way?